Do not quote single shell argument
Only a single argument is passed to the shell, then it is treated as a
shell script which should be passed in its entirety to the shell.  This
allows shell constructs such as pipes to work correctly.

Given the complexity around shell argument parsing, I'm not convinced
anymore that it makes sense for leader to spawn user commands in a shell
automatically.