Add support for fish
With the introduction of another shell, it made sense to introduce
an interface for shell-specific functionality and provide an
 implementation for each shell

Querying fish for the current command line using the `commandline`
builtin only works when fish is run interactively.  Using
github.com/kr/pty to connect fish to a pty failed due to an
unspecified input/output error.

Instead of running `commandline` from `leader`, the fish wrapper
function invokes those commands and saves the result in two
environment variables that are then accessed by `leader` to obtain the
necessary information about whether the cursor is inside a quoted
string.